backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
python-chess
openai
anthropic
python-dotenv
numpy
opencv-python
pytest
pytest-mock

# ETL pipeline dependencies
python-docx
PyMuPDF
Pillow
weaviate-client
langgraph
langchain
sentence-transformers

# Speech and audio
whisper
elevenlabs
sounddevice

# Monitoring and logging
rich

uvicorn
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0
